Move miscellaneous forms into separate module …to prevent views module growing too large
